Moved all the files to the new structure: - server/ = server binary - client/ = client library - core/crates/ = shared logic
