+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [ "crates/*", "server" ]
+members = [ "client", "crates/*", "server" ]
 
 [profile.release]
 opt-level = 3
