+++ client/Cargo.toml
@@ -1,18 +1,8 @@
 [package]
-name = "ya-relay-server"
+name = "ya-relay-client"
 version = "0.1.0"
 authors = ["Golem Factory <contact@golem.network>"]
 edition = "2018"
-
-
-[lib]
-name = "ya_relay_server"
-path = "src/lib.rs"
-
-[[bin]]
-name = "ya-relay-server"
-path = "src/main.rs"
-
 
 [dependencies]
 ya-relay-stack = { path = "../crates/stack" }
@@ -20,21 +10,15 @@ ya-relay-proto = { path = "../crates/proto" }
 ya-relay-core = { path = "../crates/core" }
 ya-client-model = "0.3"
 
-actix-rt = "1.1"
 anyhow = "1.0"
-chrono = "0.4"
-dotenv = "0.15"
+derive_more = "0.99"
 env_logger = { version = "0.8", default-features = false }
 futures = "0.3"
-itertools = "0.10.1"
 log = "0.4"
-rand = { version = "0.8", features = ["std"] }
-structopt = "0.3"
-thiserror = "1.0"
 tokio = { version = "0.2", features = ["tcp", "udp", "net", "sync", "macros", "rt-core", "stream", "time"] }
-tokio-util = { version = "0.3", features = ["codec"] }
 url = "2.1"
 
 [dev-dependencies]
-serial_test = { git = "https://github.com/tworec/serial_test.git", branch = "actix_rt_test"}
-ya-relay-client = { path = "../client" }
+actix-rt = "1.1"
+dotenv = "0.15"
+structopt = "0.3"
