+++ client/examples/client.rs
@@ -2,9 +2,9 @@ use env_logger;
 use structopt::{clap, StructOpt};
 
 use ya_client_model::NodeId;
-use ya_relay_server::crypto::FallbackCryptoProvider;
-use ya_relay_server::testing::key::{load_or_generate, Protected};
-use ya_relay_server::testing::ClientBuilder;
+use ya_relay_client::ClientBuilder;
+use ya_relay_core::crypto::FallbackCryptoProvider;
+use ya_relay_core::key::{load_or_generate, Protected};
 
 #[derive(StructOpt)]
 #[structopt(about = "NET Client")]
