+++ client/src/lib.rs
@@ -1,0 +1,4 @@
+mod client;
+mod dispatch;
+
+pub use client::{Client, ClientBuilder, ForwardReceiver};
