+++ crates/core/src/challenge.rs
