+++ crates/core/src/session.rs
