+++ server/src/testing.rs
@@ -1,8 +1,1 @@
-pub mod client;
-mod dispatch;
-pub mod key;
 pub mod server;
-
-pub use crate::server::Server;
-pub use crate::testing::client::Client;
-pub use crate::testing::client::ClientBuilder;
