+++ server/tests/test_init_session.rs
@@ -2,15 +2,15 @@ use std::convert::TryFrom;
 
 use std::ops::DerefMut;
 use ya_client_model::NodeId;
+use ya_relay_client::ClientBuilder;
+use ya_relay_core::session::SessionId;
 use ya_relay_proto::proto;
 use ya_relay_server::testing::server::init_test_server;
-use ya_relay_server::testing::ClientBuilder;
-use ya_relay_server::SessionId;
 
 #[serial_test::serial]
 async fn test_query_self_node_info() -> anyhow::Result<()> {
     let wrapper = init_test_server().await.unwrap();
-    let client = ClientBuilder::from_server(&wrapper.server)
+    let client = ClientBuilder::from_url(wrapper.server.inner.url.clone())
         .build()
         .await
         .unwrap();
@@ -42,7 +42,7 @@ async fn test_query_self_node_info() -> anyhow::Result<()> {
 #[serial_test::serial]
 async fn test_request_with_invalid_session() -> anyhow::Result<()> {
     let wrapper = init_test_server().await.unwrap();
-    let client = ClientBuilder::from_server(&wrapper.server)
+    let client = ClientBuilder::from_url(wrapper.server.inner.url.clone())
         .connect()
         .build()
         .await
@@ -75,12 +75,12 @@ async fn test_request_with_invalid_session() -> anyhow::Result<()> {
 #[serial_test::serial]
 async fn test_query_other_node_info() -> anyhow::Result<()> {
     let wrapper = init_test_server().await.unwrap();
-    let client1 = ClientBuilder::from_server(&wrapper.server)
+    let client1 = ClientBuilder::from_url(wrapper.server.inner.url.clone())
         .connect()
         .build()
         .await
         .unwrap();
-    let client2 = ClientBuilder::from_server(&wrapper.server)
+    let client2 = ClientBuilder::from_url(wrapper.server.inner.url.clone())
         .connect()
         .build()
         .await
@@ -103,7 +103,7 @@ async fn test_query_other_node_info() -> anyhow::Result<()> {
 #[serial_test::serial]
 async fn test_close_session() -> anyhow::Result<()> {
     let wrapper = init_test_server().await.unwrap();
-    let client = ClientBuilder::from_server(&wrapper.server)
+    let client = ClientBuilder::from_url(wrapper.server.inner.url.clone())
         .connect()
         .build()
         .await
